+++ semcore/tooltip/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [6.17.0] - 2024-01-19
+
+### Changed
+
+- Tooltip focus looping is disabled by default.
+
 ## [6.16.0] - 2024-01-19
 
 ### Changed